+++ webapp-backend/src/main/java/org/oransc/ric/portal/dashboard/config/XappManagerMockConfiguration.java
@@ -66,14 +66,10 @@ public class XappManagerMockConfiguration {
 		Xapp dc = new Xapp().name("Dual Connectivity").version("v2").status(StatusEnum.DELETED);
 		dc.addInstancesItem(
 				new XappInstance().name("def0-6789").ip("6.7.8.9").port(300).status(XappInstance.StatusEnum.COMPLETED));
-		Xapp pc = new Xapp().name("Pendulum Control").version("v1").status(StatusEnum.DEPLOYED);
-		pc.addInstancesItem(
-				new XappInstance().name("abcd-1234").ip("1.2.3.4").port(100).status(XappInstance.StatusEnum.RUNNING));
 		allXapps = new AllXapps();
 		allXapps.add(ac);
 		allXapps.add(an);
 		allXapps.add(dc);
-		allXapps.add(pc);
 	}
 
 	@Bean