+++ docs/release-notes.rst
@@ -22,6 +22,7 @@ RIC Dashboard Release Notes
 
 Version 1.0.2, 9 May 2019
 -------------------------
+* Update A1 mediator client to version 0.4.0
 * Add E2 response message with timestamp and status code
 
 Version 1.0.1, 6 May 2019