+++ docs/release-notes.rst
@@ -23,6 +23,7 @@ RIC Dashboard Release Notes
 Version 1.0.2, 9 May 2019
 -------------------------
 * Add E2 response message with timestamp and status code
+* Fetch xAPP instance status information from xAPP Manager and display in dashboard
 
 Version 1.0.1, 6 May 2019
 -------------------------