+++ docs/release-notes.rst
@@ -20,11 +20,12 @@
 RIC Dashboard Release Notes
 ===========================
 
-Version 1.0.2, 9 May 2019
--------------------------
+Version 1.0.2, 10 May 2019
+--------------------------
 * Update A1 mediator client to version 0.4.0
 * Add E2 response message with timestamp and status code
 * Fetch xAPP instance status information from xAPP Manager and display in dashboard
+* Allow the user to initiate an E2 (X2) connection between RIC and gNB/eNB
 
 Version 1.0.1, 6 May 2019
 -------------------------